educational_bot/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-cors-headers==4.3.0
django-environ==0.10.0
django-filter==23.3
django-jet-reboot==1.3.3
django-modeltranslation==0.18.10
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
executing==2.0.1
firebase-admin==5.2.0
frozenlist==1.3.3
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==21.1.0
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
icecream==2.1.3
idna==3.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
magic-filter==1.0.9
Markdown==3.5
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
numpy==1.25.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==3.7.0
proto-plus==1.22.2
protobuf==4.23.2
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
spyne==2.14.0
sqlparse==0.4.4
suds==1.1.2
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.6.0
xmltodict==0.13.0
yarl==1.9.2
zeep==4.2.1
